SERVICES-1535: move fees collector transactions to own service
