+++ src/modules/fees-collector/services/fees-collector.service.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import {
     FeesCollectorModel,
-    FeesCollectorTransactionModel,
     UserEntryFeesCollectorModel,
 } from '../models/fees-collector.model';
 import { EsdtTokenPayment } from '../../../models/esdtTokenPayment.model';
@@ -9,10 +8,7 @@ import {
     GlobalInfoByWeekModel,
     UserInfoByWeekModel,
 } from '../../../submodules/weekly-rewards-splitting/models/weekly-rewards-splitting.model';
-import { TransactionModel } from '../../../models/transaction.model';
-import { constantsConfig, mxConfig, gasConfig } from '../../../config';
-import { MXProxyService } from '../../../services/multiversx-communication/mx.proxy.service';
-import { Address, AddressValue } from '@multiversx/sdk-core';
+import { constantsConfig } from '../../../config';
 import BigNumber from 'bignumber.js';
 import { WeekTimekeepingModel } from 'src/submodules/week-timekeeping/models/week-timekeeping.model';
 import { FeesCollectorAbiService } from './fees-collector.abi.service';
@@ -27,53 +23,7 @@ export class FeesCollectorService {
         private readonly feesCollectorCompute: FeesCollectorComputeService,
         private readonly weekTimekeepingAbi: WeekTimekeepingAbiService,
         private readonly weeklyRewardsSplittingAbi: WeeklyRewardsSplittingAbiService,
-        private readonly mxProxy: MXProxyService,
     ) {}
-
-    async claimRewardsBatch(
-        scAddress: string,
-        userAddress: string,
-    ): Promise<FeesCollectorTransactionModel> {
-        const currentWeek = await this.weekTimekeepingAbi.currentWeek(
-            scAddress,
-        );
-        const lastActiveWeekForUser =
-            await this.weeklyRewardsSplittingAbi.lastActiveWeekForUser(
-                scAddress,
-                userAddress,
-            );
-        const weekToClaim = Math.min(
-            constantsConfig.USER_MAX_CLAIM_WEEKS,
-            currentWeek - lastActiveWeekForUser,
-        );
-        const transaction = await this.claimRewards(
-            userAddress,
-            weekToClaim * gasConfig.feesCollector.claimRewardsPerWeek +
-                gasConfig.feesCollector.baseClaimRewards,
-        );
-        const claimTransaction = new FeesCollectorTransactionModel({
-            transaction: transaction,
-            count: 0,
-        });
-        if (lastActiveWeekForUser === 0) return claimTransaction;
-        if (lastActiveWeekForUser >= currentWeek) return claimTransaction;
-
-        claimTransaction.count = 1;
-        return claimTransaction;
-    }
-
-    async claimRewards(
-        sender: string,
-        gasLimit: number,
-    ): Promise<TransactionModel> {
-        const contract = await this.mxProxy.getFeesCollectorContract();
-        return contract.methodsExplicit
-            .claimRewards()
-            .withGasLimit(gasLimit)
-            .withChainID(mxConfig.chainID)
-            .buildTransaction()
-            .toPlainObject();
-    }
 
     async getAccumulatedFees(
         scAddress: string,
@@ -206,18 +156,6 @@ export class FeesCollectorService {
             );
         }
         return modelsList;
-    }
-
-    async updateEnergyForUser(userAddress: string): Promise<TransactionModel> {
-        const contract = await this.mxProxy.getFeesCollectorContract();
-        return contract.methodsExplicit
-            .updateEnergyForUser([
-                new AddressValue(Address.fromString(userAddress)),
-            ])
-            .withGasLimit(gasConfig.feesCollector.updateEnergyForUser)
-            .withChainID(mxConfig.chainID)
-            .buildTransaction()
-            .toPlainObject();
     }
 
     async getUserAccumulatedRewards(
