+++ src/modules/fees-collector/services/fees-collector.transaction.service.ts
@@ -1,0 +1,75 @@
+import { Injectable } from '@nestjs/common';
+import { MXProxyService } from 'src/services/multiversx-communication/mx.proxy.service';
+import { FeesCollectorTransactionModel } from '../models/fees-collector.model';
+import { WeekTimekeepingAbiService } from 'src/submodules/week-timekeeping/services/week-timekeeping.abi.service';
+import { WeeklyRewardsSplittingAbiService } from 'src/submodules/weekly-rewards-splitting/services/weekly-rewards-splitting.abi.service';
+import { constantsConfig, gasConfig, mxConfig } from 'src/config';
+import { TransactionModel } from 'src/models/transaction.model';
+import { Address, AddressValue } from '@multiversx/sdk-core/out';
+
+@Injectable()
+export class FeesCollectorTransactionService {
+    constructor(
+        private readonly weekTimekeepingAbi: WeekTimekeepingAbiService,
+        private readonly weeklyRewardsSplittingAbi: WeeklyRewardsSplittingAbiService,
+        private readonly mxProxy: MXProxyService,
+    ) {}
+
+    async claimRewardsBatch(
+        scAddress: string,
+        userAddress: string,
+    ): Promise<FeesCollectorTransactionModel> {
+        const [currentWeek, lastActiveWeekForUser] = await Promise.all([
+            this.weekTimekeepingAbi.currentWeek(scAddress),
+            this.weeklyRewardsSplittingAbi.lastActiveWeekForUser(
+                scAddress,
+                userAddress,
+            ),
+        ]);
+
+        const weekToClaim = Math.min(
+            constantsConfig.USER_MAX_CLAIM_WEEKS,
+            currentWeek - lastActiveWeekForUser,
+        );
+        const transaction = await this.claimRewards(
+            weekToClaim * gasConfig.feesCollector.claimRewardsPerWeek +
+                gasConfig.feesCollector.baseClaimRewards,
+        );
+        const claimTransaction = new FeesCollectorTransactionModel({
+            transaction: transaction,
+            count: 0,
+        });
+
+        if (
+            lastActiveWeekForUser === 0 ||
+            lastActiveWeekForUser >= currentWeek
+        ) {
+            return claimTransaction;
+        }
+
+        claimTransaction.count = 1;
+        return claimTransaction;
+    }
+
+    async claimRewards(gasLimit: number): Promise<TransactionModel> {
+        const contract = await this.mxProxy.getFeesCollectorContract();
+        return contract.methodsExplicit
+            .claimRewards()
+            .withGasLimit(gasLimit)
+            .withChainID(mxConfig.chainID)
+            .buildTransaction()
+            .toPlainObject();
+    }
+
+    async updateEnergyForUser(userAddress: string): Promise<TransactionModel> {
+        const contract = await this.mxProxy.getFeesCollectorContract();
+        return contract.methodsExplicit
+            .updateEnergyForUser([
+                new AddressValue(Address.fromString(userAddress)),
+            ])
+            .withGasLimit(gasConfig.feesCollector.updateEnergyForUser)
+            .withChainID(mxConfig.chainID)
+            .buildTransaction()
+            .toPlainObject();
+    }
+}
